feat: class methods can be prioritized over others in the API Documentation page IHS-205
