+++ docs/docs_generation/content_gen_methods/mdx/__init__.py
@@ -1,8 +1,12 @@
 from __future__ import annotations
 
 from .mdx_code_doc import MdxCodeDocumentation, MdxFile
+from .mdx_ordered_section import OrderedMdxSection
+from .mdx_section import MdxSection
 
 __all__ = [
     "MdxCodeDocumentation",
     "MdxFile",
+    "MdxSection",
+    "OrderedMdxSection",
 ]
